requirements.txt
httpx>=0.27.0
rapidfuzz>=3.9.0
pyyaml>=6.0
rich>=13.7.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
twilio>=9.0.0
cryptography>=42.0.0
anthropic>=0.40.0
